+++ components-ui/app/src/test/playwright/tests/board/boardComponents/boardTable.spec.ts
@@ -5,25 +5,21 @@
 import { expect, test } from "@playwright/test";
 import {
   addComponent,
-  setBoard,
+  createPopulatedBoardAndGoToPage,
   setAuth,
   deleteComponent,
   exitEditMode,
   enterEditMode,
   clickUtilityBarButton,
-  createBoardId,
   isInHeadlessMode,
   configureTable,
 } from "../../helpers";
 
-let BOARD_ID: string;
-
 test.use({ headless: isInHeadlessMode });
 
 test.beforeEach(async ({ page }) => {
-  BOARD_ID = createBoardId();
   await setAuth(page);
-  await setBoard(page, BOARD_ID);
+  await createPopulatedBoardAndGoToPage(page);
   await enterEditMode(page);
 });