+++ components-ui/app/src/test/playwright/tests/board/boardTranslators/boardTranslator.spec.ts
@@ -6,9 +6,9 @@ import { test, expect } from "@playwright/test";
 import {
   setAuth,
   isInHeadlessMode,
-  insertComponentToBoard,
-  insertZoneToBoard,
-  createBoardForUser,
+  createComponentInZone,
+  createZoneInView,
+  createBoardAndViewAndZone,
   createTableConfig,
 } from "../../helpers";
 import { TRANSLATOR_TEST_INPUTS } from "./boardTranslatorConfigs.config";
@@ -25,14 +25,14 @@ TRANSLATOR_TEST_INPUTS.forEach(({ zoneObjectTypes, TableFields }) => {
       const userId = JSON.parse(user || '{}').id;
 
       // Create an initial board
-      const { boardId, viewId } = await createBoardForUser({
+      const { boardId, viewId } = await createBoardAndViewAndZone({
         userId: String(userId),
       });
 
       // Create additional zones for the board
       let zoneIds: string[] = [ ];
       for (const zoneObjectType of zoneObjectTypes || []) {
-        const returnedZoneId = await insertZoneToBoard({
+        const returnedZoneId = await createZoneInView({
           userId: String(userId),
           viewId,
           title: `Zone for ${zoneObjectType}`,
@@ -48,7 +48,7 @@ TRANSLATOR_TEST_INPUTS.forEach(({ zoneObjectTypes, TableFields }) => {
       // For each new zone, insert a table component with the corresponding TableFields
       for (const [index, newZoneId] of zoneIds.entries()) {
         const field = TableFields[zoneObjectTypes[index]];
-        await insertComponentToBoard(
+        await createComponentInZone(
           {
             userId: String(userId),
             componentTitle: `${newZoneId} Table`,