+++ CHANGELOG.md
@@ -28,6 +28,12 @@ SPDX-License-Identifier: MIT
 
 - Updated table download modal to use dataSource
 
+## tol-ui `4.0.6`
+
+28-07-2026
+
+- Updated table download modal to use dataSource
+
 ## tol-ui `4.0.5`
 
 27-07-2026